플레이어가 획득한 통합 코인 수를 얻는 Getter의 잘못된 반환형 수정

+++ Game/Inc/Game/GameManager.h
@@ -55,7 +55,7 @@ public:
 
 	/** 통합 게임 플레이 기록에 대한 Getter 입니다. */
 	float GetTotalPlayTime() const { return totalPlayTime_; }
-	float GetTotalCoin() const { return totalCoin_; };
+	int32_t GetTotalCoin() const { return totalCoin_; };
 	int32_t GetTotalUseSkill() const { return totalUseSkill_; }
 	float GetTotalHealHp() const { return totalHealHp_; }
 	float GetTotalLostHp() const { return totalLostHp_; }
